lint

+++ src/CompoundV3YieldOracle.sol
@@ -38,7 +38,7 @@ contract CompoundV3YieldOracle is ICompoundV3YieldOracle, Keep3rV2Job {
 
     IERC20[] public tokenRefreshList;
 
-    mapping(address => uint) private tokenToTokenRefreshListIndex;
+    mapping(address => uint256) private tokenToTokenRefreshListIndex;
 
     constructor(address _keep3r) {
         admin = msg.sender;
